Rename PA folder to CrossPlatform
