+++ Source/Spatializer/CrossPlatform/AudioPluginMicrosoftAcoustics.def
@@ -1,0 +1,2 @@
+LIBRARY
+EXPORTS
