+++ Source/Spatializer/CrossPlatform/PluginList.h
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+DECLARE_EFFECT("Project Acoustics", Spatializer)
+DECLARE_EFFECT("Project Acoustics Mixer", SpatialMixer)
