--- Source/Spatializer/ProjectAcoustics/AudioPluginMicrosoftAcoustics.def
@@ -1,2 +1,0 @@
-LIBRARY
-EXPORTS
