--- Source/Spatializer/ProjectAcoustics/PluginList.h
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-DECLARE_EFFECT("Project Acoustics", Spatializer)
-DECLARE_EFFECT("Project Acoustics Mixer", SpatialMixer)
